ref: organized asserts and mockito
