+++ src/api/models/delegate.rs
@@ -2,6 +2,7 @@ use serde::Deserialize;
 use serde::Serialize;
 
 use crate::api::models::timestamp::Timestamp;
+use crate::common::deserialize_u64_as_number_or_string;
 
 #[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
@@ -9,7 +10,7 @@ pub struct Delegate {
     pub username: String,
     pub address: String,
     pub public_key: String,
-    pub votes: u64,
+    pub votes: String,
     pub rank: u32,
     pub blocks: Blocks,
     pub production: Production,
@@ -19,7 +20,6 @@ pub struct Delegate {
 #[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
 pub struct Blocks {
     pub produced: u64,
-    pub missed: u64,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub last: Option<Last>,
 }
@@ -27,18 +27,21 @@ pub struct Blocks {
 #[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
 pub struct Production {
     pub approval: f64,
-    pub productivity: f64,
 }
 
 #[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
 pub struct DelegateForged {
+    #[serde(deserialize_with = "deserialize_u64_as_number_or_string")]
     pub rewards: u64,
+    #[serde(deserialize_with = "deserialize_u64_as_number_or_string")]
     pub fees: u64,
+    #[serde(deserialize_with = "deserialize_u64_as_number_or_string")]
     pub total: u64,
 }
 
 #[derive(Clone, Debug, Default, PartialEq, Deserialize, Serialize)]
 pub struct Last {
     pub id: String,
+    pub height: u64,
     pub timestamp: Timestamp,
 }
