+++ tests/api/transactions_test.rs
@@ -1,9 +1,10 @@
-use crate::common::{
-    assert_meta, assert_transaction_data, mock_client, mock_http_request, mock_post_request,
-};
+use std::collections::HashMap;
+
 use serde_json::from_str;
 use serde_json::Value;
-use std::collections::HashMap;
+
+use crate::utils::assert_helpers::{assert_meta, assert_transaction_data};
+use crate::utils::mockito_helpers::{mock_client, mock_http_request, mock_post_request};
 
 #[test]
 fn test_all() {
