+++ tests/api/votes_test.rs
@@ -1,6 +1,10 @@
 use serde_json::from_str;
 
-use crate::common::*;
+
+use crate::utils::assert_helpers::{assert_transaction_data, assert_meta};
+use crate::utils::mockito_helpers::{mock_client, mock_http_request};
+
+
 use arkecosystem_client::api::models::asset::Asset;
 use arkecosystem_client::api::models::transaction::Transaction;
 use serde_json::Value;
