+++ tests/lib.rs
@@ -1,3 +1,3 @@
 mod api;
-mod common;
+mod utils;
 mod connection;
